etong 4 na file, pa TAKE NOTE (UPDATE NYO REN SQL NYO)
this may fail... ang prob ko here is:

the generated otp will be send to the email of the user (the generated otp is saved to the database and that otp is the same to what it will be emailed to the user). when the link is clicked, the user will enter that otp and also its new password. if the user type their otp and new password, if the otp is incorrect, it will alert wrong otp.

another thing that i want to happen is, there would be an expiration of generated otp for 3mins. so the flow is, when the user generate otp, that otp will be send to the email and be save to the database, but after 3mins, the otp will automatically be deleted in the database

(what i want to happen when the user click the link from their email. , the user will enter that otp and also its new password. if the user type their otp and new password, if the otp is incorrect, it will alert wrong otp. make sure that what the user receive from their email is matched to the database so that the new password wont be updated to the database because otp is not matched to the database)